Fully support the variable declaration and assignment with record referencing
