+++ src/parser/index.ts
@@ -4,8 +4,8 @@ import { parseIdentifier } from './identifier';
 import { parsePrimitive } from './primitives';
 import { parseTypeLiteral } from './type-literal';
 import { parseListLiteral } from './composite-literals';
-import { parseRecordDeclaration } from './record';
-import { parseRecordReferenceExpr } from './record-reference';
+import { parseRecordDeclaration } from './record/declaration';
+import { parseRecordReferenceExpr } from './record/reference';
 import { parseVarDeclaration } from './variable-declaration';
 import { parseFunctionDeclaration, parseMethodDeclaration } from './function';
 import { parseConditionalExpr } from './condition';
