+++ src/samples/record/record-assignment.ts
@@ -5,6 +5,7 @@ import { DataType as DT, Scope } from '../../parser/utils';
 const program = `\
 age = maxwell->age
 mutable name = maxwell->name
+name = maxwell->name
 `;
 
 const tokens: Array<Token> = [
@@ -16,6 +17,13 @@ const tokens: Array<Token> = [
   { type: 'newline', value: '\n' },
 
   { type: 'keyword', value: 'mutable' },
+  { type: 'ident', value: 'name' },
+  { type: 'eq', value: '=' },
+  { type: 'ident', value: 'maxwell' },
+  { type: 'ref', value: '->' },
+  { type: 'ident', value: 'name' },
+  { type: 'newline', value: '\n' },
+
   { type: 'ident', value: 'name' },
   { type: 'eq', value: '=' },
   { type: 'ident', value: 'maxwell' },
@@ -47,14 +55,26 @@ const ast: AST = [
       return: DT.Str
     },
   },
+  {
+    type: 'VarAssignmentExpr',
+    return: DT.Void,
+    expr1: Var('name', DT.Str),
+    expr2: {
+      type: 'RecordReferenceExpr',
+      recordExpr: Var('maxwell', new DT('UserInfo')),
+      property: 'name',
+      return: DT.Str
+    },
+  },
 ];
 
 const compiled = `\
 const age = maxwell.age;
 let name = maxwell.name;
+name = maxwell.name;
 `;
 
-const minified = 'const age=maxwell.age;let name=maxwell.name;';
+const minified = 'const age=maxwell.age;let name=maxwell.name;name=maxwell.name;';
 
 const scope = () => {
   const s = new Scope();
